Update to 9.1.0

+++ SpellWork/DBC/Structures/SpellXSpellVisualEntry.cs
@@ -9,7 +9,7 @@ namespace SpellWork.DBC.Structures
         public byte DifficultyID;
         public uint SpellVisualID;
         public float Probability;
-        public byte Priority;
+        public int Priority;
         public int SpellIconFileID;
         public int ActiveIconFileID;
         public ushort ViewerUnitConditionID;
